requirements.txt
arabic-reshaper==2.1.3
wordcloud==1.8.1
python-bidi==0.4.2
hazm==0.7.0
loguru==0.6.0
